+++ paperscraper/pdf/__init__.py
@@ -1,0 +1,1 @@
+from .pdf import load_api_keys, save_pdf, save_pdf_from_dump  # noqa
